pyproject.toml
[tool.poetry]

name = "memegen"
version = "0.1.0"
description = "A meme generator API"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

packages = [
    { include = "app" },
    { include = "app/**/*.py" }
]

[tool.poetry.dependencies]

python = "^3.12"

# Sanic
sanic = "^24.6.0"
sanic-ext = "^23.12.0"

# Markdown
markdown = "~3.5.1"
pymdown-extensions = "^10.0"
pygments = "^2.15.0"

# Images
pillow = "^10.4.0"
pilmoji = { git = "https://github.com/jay3332/pilmoji", branch = "2.0" }
spongemock = "~0.3.4"

# Utilities
aiocache = "~0.11.1"
aiohttp = "^3.10.11"
anyio = "^3.6.2"
datafiles = "^2.2.3"
furl = "^2.1.3"
webp = "~0.3"

# Production
bugsnag = "^4.6"
gunicorn = "^22.0"
uvicorn = "~0.23.2"
locust = { version = "^2.12", optional = true }

# Embeddings
chromadb = "^0.5.23"
sentence-transformers = "^3.3.1"
pyyaml = "^6.0.2"
beautifulsoup4 = "^4.12.3"
httpx = "^0.28.1"
openai = "^1.58.1"

[tool.poetry.group.dev.dependencies]

# Formatting
isort = "^5.12"
black = "^24.3"

# Type Checking
mypy = "^1.4.1"
autoflake = "^2.1"
types-aiofiles = "*"
types-pillow = "*"
types-requests = "*"
types-setuptools = "*"

# Testing
pytest = "^7.1.3"
pytest-asyncio = "*"
pytest-describe = "^2.0"
pytest-expecter = "^3.0"
pytest-random = "*"
pytest-cov = "^4.1"
sanic-testing = "*"
aioresponses = "*"

# Coverage
coveragespace = "^6.1"

# Documentation
mkdocs = "^1.2.3"

# Utilities
honcho = "^1.1"
ipdb = "*"
rope = "^0.18"
watchdog = { version = "^2.2", extras = ["watchmedo"] }
ipykernel = "^6.29.5"

[tool.poetry.extras]

testing = ["locust"]

[tool.black]

quiet = true

[tool.isort]

profile = "black"

[tool.mypy]

ignore_missing_imports = true
no_implicit_optional = true
check_untyped_defs = true

plugins = "datafiles.plugins:mypy"

cache_dir = ".cache/mypy/"

[tool.pytest.ini_options]

testpaths = "app"

addopts = """
--no-header \
--random \
--show-capture=log \
--cov=app --cov-report=term-missing:skip-covered --cov-report=html:.cache/htmlcov --no-cov-on-fail
"""

cache_dir = ".cache/pytest/"

markers = "slow"

asyncio_mode = "strict"

[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
